+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.16
 
 - Editing a headline, when inserting or renaming a node, can now be interrupted by most other keyboard shortcuts and commands.
+- Fixed minibuffer history so that each opened Leo document has its own minibuffer command history: Requires Leo 6.7.5
 
 ## 1.0.15